Cursor should be at the end of the line at recall
Hitting UP to recall a previous command from history should position the
cursor at the end of the line.  Apparently this line in /etc/zsh/zshrc:

    [[ "$terminfo[kcuu1]" == "^[O"* ]] && \
      bindkey -M viins "${terminfo[kcuu1]/O/[}" vi-up-line-or-history

is what causes the cursor to reset to the start of the line.

We can reliably avoid the problem by not pulling in the global rc.
Anything I miss from the global rc can be reintegrated in to my rc.