Aborted attempt at switching zsh cursor styles
tmux and iTerm2 both support (different) terminfo extensions to change the shape
of the cursor (block, vertical line, underline).  It would be nice to map a
different cursor shape to the insert and command modes so that I could tell, at
a glance, which mode I'm in.

Unfortunately, this doesn't seem to work properly when running Vim inside mosh.
Leave this snippet here as a reminder to try again some time in the future.