zsh: noglobalrcs as early as possible
~/.zshenv is read for all shells.  This file is read before ~/.zshrc.
This is the place to 'setopt noglobalrcs' if we want to reliably avoid
undesirable and unpredictable behaviour from global rcs.

This change was made in response to a change that Apple made to OS X
10.11 (El Capitan).  I belatedly upgraded to this OS X release today
only to find that my $PATH had again been inverted by path_helper(8).
On El Capitan, /etc/zprofile invokes path_helper(8) -- zsh sequences a
read from this file between ~/.zshenv and ~/.zshrc.