vim: Uninstall vim-bracketed-paste
This plugin saps a whopping ~2.8 sec from vim's startup time.

This plugin may be occasionally useful, though having to wait such a
long time for vim to load is always frustrating.

tpope/vim-unimpaired, which I already have installed, provides a
convenient 'yo' mapping in normal mode that enables paste for the
duration of the insert.  I expect that mapping will be a satisfactory
replacement for the plugin removed in this commit.