Remove ~/.gitk from version control
The contents of this file are subject to significant churn.  (The file
is dirtied almost as often as I start gitk, which admittedly doesn't
happen _that_ often, but often enough to be annoying.)

Besides, I gain little from subjecting this file to version control.  I
use gitk on just one machine, and don't particularly care if its
configuration is lost.