Revert "zsh: Reset git commit date on 'amend'"
This reverts commit 84d050f353f9db8a5438654ae9d795890c2bed1e.

On second thought, it was not a great idea to enable this behaviour by
default.  I occasionally need to make trivial amendments to several
commits in a chain undergoing review.  In such cases, it is preferable
to retain the original commit time.  Rewriting the timestamps would give
the impression that I am capable of bursting several hundreds of SLOC
per minute.  ;-)