tmux: Only supply -c against newer tmux releases
No matter which version of tmux is installed, new panes should be
created with a working directory set to the working directory of the
process running in the active pane.

tmux releases prior to 1.9 did this out of the box.

tmux releases since 1.9 require the -c flag (see commit 0feda60c).

The -c flag was not available in previous tmux releases.  0feda60c broke
.tmux.conf for, e.g., tmux 1.6, which I still use in places.