zsh: avoid read builtin in .zshenv
For reasons I cannot satisfactorily explain right now, it seems it is
best to avoid the read builtin during non-interactive shell startup.
Prior to this commit, and only when zsh was executed as a non-
interactive shell in a new process group, zsh would suspend itself
whenever it reached this line.

I first noticed this problem after installing fzf 0.19.0, which changed
the default shell for FZF_DEFAULT_COMMAND from sh to the value of SHELL.
fzf has always (or almost always) exec'd its children in new process
groups.

    % zsh --version
    zsh 5.3 (x86_64-apple-darwin18.0)

Directories with spaces in them seem to be the cause all manner of chaos
when inserted into PATH, so skip them.  No need for sh-style path
hardening idioms.