vim: fix formatting of unordered lists in asciidoc
I don't always wrap asciidoc, but when I do, it's nice to have working
autoformatting of unordered lists.

Collections like [A-Za-z] were replaced with atoms like \a because the
latter perform better.  See :help /\]

comments defaults:

   s1:/*,mb:*,ex:*/,://,b:#,:%,:XCOMM,n:>,fb:-

The default was somehow interfering with formatlistpat.  Maybe the mb:*?
Neutralising this option is enough to allow flp to work -- but we may as
well set it to a legitimate asciidoc comment marker while here.

commentstring is used by tpope/vim-commentary to toggle comments.