zsh: NO_COLOR=1
Backtraces from py3.13+ are unreadable.

This is zoomer brainrot.
https://no-color.org/

NO_COLOR is wrong but it is too hard to fight the stupidity.
People who want colour should need to turn it on --
not the other way around.

A little bit of colour might, in very limited cases,
help draw attention to an important detail.
What is the point if you print every line in colour?
This is shit user interface design.